Bump version 'y' to fix version number.
This commit aims to fix the bad version numbering.

- in the RELEASE_3_8 branch, a specific merge of the master branch,
caused bad version numbers which is preventing the user to get past the
pre receive hook.

- Since people have already been using the current version, fixing the
previous version number mistakes is not an option. The only way forward
is to upgrade the version number.

- the email
https://stat.ethz.ch/pipermail/bioc-devel/2019-January/014609.html lists
all the mistakes and explains in detail the issues.